+++ lib/src/fcm_interop.dart
@@ -37,7 +37,7 @@ class Messaging {
 class Promise<T> {
   external Promise(void executor(void resolve(T result), Function reject));
 
-  external Promise<T> then(void onFulfilled(T result), [Function onRejected]);
+  external Promise<T> then(void onFulfilled(T result), [Function? onRejected]);
 }
 
 @anonymous