+++ lib/src/stringify_interop.dart
@@ -5,4 +5,4 @@ import 'package:js/js.dart';
 
 // Calls invoke JavaScript `JSON.stringify(obj)`.
 @JS('JSON.stringify')
-external String stringify(Object obj);
+external String stringify(dynamic obj);